+++ lib/sshez/exec.rb
@@ -1,9 +1,9 @@
 require 'forwardable'
-module Sshez  
+module Sshez
   #
-  # handles all the updates to the .ssh/config file
+  # handles all the ssh commands and updates to the .ssh/config file
   #
-  class FileManager
+  class Exec
     extend Forwardable
     FILE_PATH = File.expand_path('~') + '/.ssh/config'
     PRINTER = PrintingManager.instance
@@ -37,12 +37,26 @@ module Sshez
 
     private
     #
-    # append an aliase for the given user@host with the options passed
+    # connects to host using alias
     #
-    def add(aliase_name, user, host, options)
+    def connect(alias_name, options)
+      file = File.open(FILE_PATH, 'r')
+      servers = all_hosts_in(file)
+      if servers.include?alias_name
+        PRINTER.verbose_print "Connecting to #{alias_name}"
+        exec "ssh #{alias_name}"
+      else
+        PRINTER.print "Could not find host `#{alias_name}`"
+      end
+    end
+
+    #
+    # append an alias for the given user@host with the options passed
+    #
+    def add(alias_name, user, host, options)
       begin
         PRINTER.verbose_print "Adding\n"
-        config_append = form(aliase_name, user, host, options)
+        config_append = form(alias_name, user, host, options)
         PRINTER.verbose_print config_append
         unless options.test
           file = File.open(FILE_PATH, 'a+')
@@ -50,56 +64,53 @@ module Sshez
           file.close
 
           # causes a bug in fedore if permission was not updated to 0600
-          File.chmod(0600, FILE_PATH) 
+          File.chmod(0600, FILE_PATH)
           # system "chmod 600 #{FILE_PATH}"
         end
-      rescue 
+      rescue
         return permission_error
       end
       PRINTER.verbose_print "to #{FILE_PATH}"
-      PRINTER.print "Successfully added `#{aliase_name}` as an alias for `#{user}@#{host}`"
-      PRINTER.print "try ssh #{aliase_name}"
+      PRINTER.print "Successfully added `#{alias_name}` as an alias for `#{user}@#{host}`"
+      PRINTER.print "Try sshez connect #{alias_name}"
 
       finish_exec
-    end # append(aliase_name, user, host, options)
+    end # add(alias_name, user, host, options)
 
     #
     # returns the text that will be added to the config file
     #
-    def form(aliase_name, user, host, options)
+    def form(alias_name, user, host, options)
       retuned = "\n"
-      retuned += "Host #{aliase_name}\n"
+      retuned += "Host #{alias_name}\n"
       retuned += "  HostName #{host}\n"
       retuned += "  User #{user}\n"
-      
+
       options.file_content.each_pair do |key, value|
         retuned += value
       end
       retuned
 
-    end # form(aliase_name, user, host, options)
+    end # form(alias_name, user, host, options)
 
     #
-    # removes an aliase from the config file (all its occurrences will be removed too)
+    # removes an alias from the config file (all its occurrences will be removed too)
     #
-    def remove(aliase_name, options)
+    def remove(alias_name, options)
       file = File.open(FILE_PATH, 'r')
-      new_file = File.open(FILE_PATH+'temp', 'w')
-      remove_alias_name(aliase_name, file, new_file)
-      file.close
-      new_file.close
+      new_file = File.open(FILE_PATH + 'temp', 'w')
+      remove_alias_name(alias_name, file, new_file)
+
       File.delete(FILE_PATH)
       File.rename(FILE_PATH + 'temp', FILE_PATH)
-
-      # causes a bug in fedore if permission was not updated to 0600
-      File.chmod(0600, FILE_PATH) 
-      # system "chmod 600 #{FILE_PATH}"
+      # Causes a bug in fedore if permission was not updated to 0600
+      File.chmod(0600, FILE_PATH)
 
       unless PRINTER.output?
-        PRINTER.print "could not find host (#{aliase_name})"
+        PRINTER.print "Could not find host `#{alias_name}`"
       end
       finish_exec
-    end # remove(aliase_name, options)
+    end # remove(alias_name, options)
 
     #
     # copies the content of the file to the new file without
@@ -108,24 +119,29 @@ module Sshez
     def remove_alias_name(alias_name, file, new_file)
       started_removing = false
       file.each do |line|
-        not_copying = line.include?("Host #{alias_name}") || started_removing
-        if not_copying
+        started_removing ||= line.include?("Host #{alias_name}")
+        if started_removing
           # I will never stop till I find another host that is not the one I'm removing
-          condition = (started_removing && line.include?('Host ') && !line.include?(alias_name))
-          PRINTER.verbose_print line unless condition
-          started_removing = !condition
+          stop_removing = (started_removing && line.include?('Host ') && !(line =~ /\b#{alias_name}\b/))
+          PRINTER.verbose_print line unless stop_removing
+          if stop_removing && started_removing
+            new_file.write(line)
+          end
+          started_removing = !stop_removing
         else
           # Everything else should be transfered safely to the other file
           new_file.write(line)
         end
       end
+      file.close
+      new_file.close
     end #remove_alias_name(alias_name, file, new_file)
 
     #
     # lists the aliases available in the config file
     #
     def list(options)
-      file = File.open(FILE_PATH, 'r')
+      file = File.open(FILE_PATH, 'a+')
       servers = all_hosts_in(file)
       file.close
       if servers.empty?
@@ -137,6 +153,11 @@ module Sshez
       finish_exec
     end # list(options)
 
+    def reset(options)
+      file = File.open(FILE_PATH, "w")
+      file.close
+    end
+
     #
     # Returns all the alias names of in the file
     #
@@ -144,7 +165,7 @@ module Sshez
       servers = []
       file.each do |line|
         if line.include?('Host ')
-          servers << line.sub('Host', '')
+          servers << line.sub('Host ', '').strip
         end
       end
       servers